--- a/appveyor.yml
+++ b/appveyor.yml
@@ -1,4 +1,4 @@
-version: 4.2.1.{build}
+version: 4.2.2.{build}
 pull_requests:
   do_not_increment_build_number: true
 skip_tags: true