```c#
        [Benchmark]
        public void Parse()
        {
            try
            {
                var parser = new Parser(new Lexer());
                parser.Parse(new Source(@"query { field1 field2 field3 }"));
            }
            catch (GraphQLSyntaxErrorException)
            {
            }
        }
```
``` ini

BenchmarkDotNet=v0.11.4, OS=Windows 10.0.17134.1130 (1803/April2018Update/Redstone4)
Intel Core i5-6200U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
Frequency=2343750 Hz, Resolution=426.6667 ns, Timer=TSC
.NET Core SDK=3.0.100
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 32bit RyuJIT
  DefaultJob : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 32bit RyuJIT
```
It should be noted that the measurement was not accurate, so attention should first be paid to the amount of allocated memory.

Before:

| Method |     Mean |     Error |    StdDev |   Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |---------:|----------:|----------:|---------:|------------:|------------:|------------:|--------------------:|
|  Parse | 1.814 us | 0.0362 us | 0.0915 us | 1.775 us |      0.7496 |           - |           - |             1.15 KB |

After:

| Method |     Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |---------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|  Parse | 1.922 us | 0.0349 us | 0.0309 us |      0.6599 |           - |           - |             1.02 KB |