@sungam3r 2019-12-09
A more substantial optimization could be made, but I did not want to change the code design and limited myself to only replacing the `new List<T>`. A good example for the most frequent savings is the directives in the queries.